📝 Add Toggle docs
