+++ packages/eds-core-react/src/components/Button/Button.stories.tsx
@@ -2,24 +2,13 @@ import { useState, useEffect } from 'react'
 import {
   Button,
   ButtonProps,
-  ToggleButtonProps,
   Icon,
   EdsProvider,
   Density,
   Progress,
 } from '../..'
 import { Story, ComponentMeta } from '@storybook/react'
-import {
-  menu,
-  add,
-  save,
-  edit,
-  copy,
-  calendar,
-  time,
-  alarm,
-  timer,
-} from '@equinor/eds-icons'
+import { menu, add, save } from '@equinor/eds-icons'
 import { Stack } from './../../../.storybook/components'
 import page from './Button.docs.mdx'
 
@@ -325,89 +314,6 @@ export const Compact: Story<ButtonProps> = () => {
   )
 }
 Compact.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-]
-
-export const ToggleExclusiveSelection: Story<ToggleButtonProps> = () => {
-  return (
-    <Button.Toggle aria-label="file actions">
-      <Button aria-label="save action">
-        <Icon data={save} title="Ghost icon save"></Icon>
-      </Button>
-      <Button aria-label="edit action">
-        <Icon data={edit} title="Ghost icon edit"></Icon>
-      </Button>
-      <Button aria-label="copy action">
-        <Icon data={copy} title="Ghost icon copy"></Icon>
-      </Button>
-    </Button.Toggle>
-  )
-}
-ToggleExclusiveSelection.storyName = 'Toggle Exclusive selection'
-ToggleExclusiveSelection.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-]
-
-export const ToggleMultiple: Story<ToggleButtonProps> = () => {
-  return (
-    <Button.Toggle multiple aria-label="date and time actions">
-      <Button aria-label="calendar">
-        <Icon data={calendar} title="Ghost icon calendar"></Icon>
-      </Button>
-      <Button aria-label="time">
-        <Icon data={time} title="Ghost icon time"></Icon>
-      </Button>
-      <Button aria-label="alarm">
-        <Icon data={alarm} title="Ghost icon alarm"></Icon>
-      </Button>
-      <Button aria-label="timer">
-        <Icon data={timer} title="Ghost icon timer"></Icon>
-      </Button>
-    </Button.Toggle>
-  )
-}
-ToggleMultiple.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-]
-
-export const ToggleControlled: Story<ToggleButtonProps> = () => {
-  const [selectedButtons, setSelectedButtons] = useState([0, 1])
-  const handleChange = (indexes: number[]) => {
-    setSelectedButtons(indexes)
-  }
-
-  return (
-    <Button.Toggle
-      multiple
-      selectedIndexes={selectedButtons}
-      onChange={handleChange}
-      aria-label="date and time actions"
-    >
-      <Button aria-label="calendar">
-        <Icon data={calendar} title="Ghost icon calendar"></Icon>
-      </Button>
-      <Button aria-label="time">
-        <Icon data={time} title="Ghost icon time"></Icon>
-      </Button>
-      <Button aria-label="alarm">
-        <Icon data={alarm} title="Ghost icon alarm"></Icon>
-      </Button>
-    </Button.Toggle>
-  )
-}
-ToggleControlled.decorators = [
   (Story) => (
     <Stack>
       <Story />
